// frontend/src/api/generated/model/ndmsNativeImportTargetRange.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * keen-pbr API
 * REST API for the keen-pbr policy-based routing daemon.
 * OpenAPI spec version: 3.0.0
 */
import type { NdmsNativeImportTargetRangePrefix } from './ndmsNativeImportTargetRangePrefix';

export interface NdmsNativeImportTargetRange {
  prefix: NdmsNativeImportTargetRangePrefix;
  /**
     * @minimum 0
     * @maximum 126
     */
  first_index: number;
  /**
     * @minimum 0
     * @maximum 126
     */
  last_index: number;
}